Feat: 챗봇 atom 생성
