+++ src/commons/Chatbot/questionsAndAnswers.ts
@@ -1,0 +1,36 @@
+export const intro: string =
+  "안녕하세요. 고투게더 여행 그룹 추천 상담봇입니다. \n 아직 여행지와 동행 그룹을 정하지 못하셨다면 설문을 통해 추천받아보세요! \n 총 10개의 문항이 있으며, 모든 항목은 중복 선택이 가능합니다. \n 응답 제출 시 SMS 또는 전화로 여행 그룹을 추천받으실 수 있습니다.";
+
+export const questions: string[] = [
+  "정확한 결과 발송을 위해 연락처를 남겨주세요. \n ex) 01012345678",
+  "이름 또는 닉네임을 입력해주세요.",
+  "함께 여행하고 싶은 연령대 그룹은?",
+  "함께 여행하고 싶은 그룹 유형은?",
+  "여행에 함께 참가하는 동행자의 유형은?",
+  "함께 여행하고 싶은 종교적 성향의 그룹은?",
+  "함께 여행하고 싶은 정치적 성향의 그룹은?",
+  "가고 싶은 여행의 테마는?",
+  "원하는 여행 시기는?",
+  "연평균 해외여행 횟수는?",
+  "설문이 종료되었습니다. 응답하신 내용을 바탕으로 최적의 그룹을 찾아 연락드리겠습니다!",
+];
+
+// 질문 순서에 따른 버튼목록
+export const answers = [
+  ["2030", "3040", "4050", "5060", "6070", "모든 연령대"],
+  ["여자끼리", "남자끼리", "자녀동반", "상관없음"],
+  ["나홀로", "친구/동료", "연인/부부", "자녀 동반 가족", "상관없음"],
+  ["하나님과 함께하는 여행", "상관없음", "부처님의 발자취를 찾아"],
+  ["보수적인 그룹", "진보적인 그룹", "상관없음"],
+  [
+    "휴양 및 힐링",
+    "문화탐방",
+    "골프여행",
+    "오지탐험",
+    "트레킹",
+    "봉사활동",
+    "성지순례",
+  ],
+  ["봄", "여름", "가을", "겨울", "아무때나"],
+  ["연 1회 이하", "연 2회", "연 3회", "연 4회", "연 5회 이상"],
+];
